api/requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
black==20.8b1
Click==7.1.2
dealer==2.1.0
decorator==4.4.2
Django==3.1.3
django-admin-list-filter-dropdown==1.0.3
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
entrypoints==0.3
flake8==3.8.4
h11==0.11.0
httptools==0.1.1
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
parso==0.7.0
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.1.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.2
python-magic==0.4.18
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.11
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.1
uvicorn==0.11.7
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
django-bootstrap4==2.3.1
django-oauth-toolkit==1.3.3
django-cors-middleware==1.5.0
django-environ==0.4.5
djangorestframework==3.12.2
django-extensions==3.0.9
drf-yasg==1.20.0
drf-writable-nested==0.6.2
pytest-django==4.1.0
pytest==6.1.2
pytest-cov==2.10.1
factory_boy==3.1.0
persisting-theory==0.2.1
pytest-env==0.6.2
